refactor: move Error to its own package
