+++ api/account/api.go
@@ -13,6 +13,7 @@ import (
 	"github.com/emerishq/demeris-api-server/api/apiutils"
 	"github.com/emerishq/demeris-api-server/api/database"
 	"github.com/emerishq/demeris-api-server/api/router/deps"
+	"github.com/emerishq/demeris-api-server/lib/apierrors"
 	"github.com/emerishq/demeris-api-server/sdkservice"
 	"github.com/emerishq/demeris-backend-models/cns"
 	"github.com/emerishq/demeris-backend-models/tracelistener"
@@ -48,7 +49,7 @@ func GetBalancesByAddress(c *gin.Context) {
 	balances, err := d.Database.Balances(address)
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"account",
 			fmt.Errorf("cannot retrieve account for address %v", address),
 			http.StatusBadRequest,
@@ -68,7 +69,7 @@ func GetBalancesByAddress(c *gin.Context) {
 
 	vd, err := verifiedDenomsMap(d.Database)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"account",
 			fmt.Errorf("cannot retrieve account for address %v", address),
 			http.StatusBadRequest,
@@ -181,7 +182,7 @@ func GetDelegationsByAddress(c *gin.Context) {
 	dl, err := d.Database.Delegations(address)
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"delegations",
 			fmt.Errorf("cannot retrieve delegations for address %v", address),
 			http.StatusBadRequest,
@@ -231,7 +232,7 @@ func GetUnbondingDelegationsByAddress(c *gin.Context) {
 	unbondings, err := d.Database.UnbondingDelegations(address)
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"unbonding delegations",
 			fmt.Errorf("cannot retrieve unbonding delegations for address %v", address),
 			http.StatusBadRequest,
@@ -284,7 +285,7 @@ func GetDelegatorRewards(c *gin.Context) {
 
 	chain, err := d.Database.Chain(chainName)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve chain with name %v", chainName),
 			http.StatusBadRequest,
@@ -305,7 +306,7 @@ func GetDelegatorRewards(c *gin.Context) {
 
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusInternalServerError,
@@ -331,7 +332,7 @@ func GetDelegatorRewards(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve delegator rewards from sdk-service"),
 			http.StatusInternalServerError,
@@ -407,7 +408,7 @@ func GetNumbersByAddress(c *gin.Context) {
 		dl, err := d.Database.Numbers(address)
 
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"numbers",
 				fmt.Errorf("cannot retrieve account/sequence numbers for address %v", address),
 				http.StatusBadRequest,
@@ -428,7 +429,7 @@ func GetNumbersByAddress(c *gin.Context) {
 
 	resp, err := fetchNumbers(dd, address)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"numbers",
 			fmt.Errorf("cannot retrieve account/sequence numbers for address %v", address),
 			http.StatusInternalServerError,
@@ -459,7 +460,7 @@ func GetUserTickets(c *gin.Context) {
 
 	tickets, err := d.Store.GetUserTickets(address)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"tickets",
 			fmt.Errorf("cannot retrieve tickets for address %v", address),
 			http.StatusBadRequest,
