+++ api/block/api.go
@@ -11,6 +11,7 @@ import (
 	"github.com/emerishq/emeris-utils/store"
 
 	"github.com/emerishq/demeris-api-server/api/router/deps"
+	"github.com/emerishq/demeris-api-server/lib/apierrors"
 	"github.com/gin-gonic/gin"
 )
 
@@ -33,7 +34,7 @@ func GetBlock(c *gin.Context) {
 
 	h := c.Query("height")
 	if h == "" {
-		e := deps.NewError(
+		e := apierrors.New(
 			"block",
 			fmt.Errorf("missing height"),
 			http.StatusBadRequest,
@@ -49,7 +50,7 @@ func GetBlock(c *gin.Context) {
 
 	hh, err := strconv.ParseInt(h, 10, 64)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"block",
 			fmt.Errorf("malformed height"),
 			http.StatusBadRequest,
@@ -71,7 +72,7 @@ func GetBlock(c *gin.Context) {
 
 	bd, err := bs.Block(hh)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"block",
 			fmt.Errorf("cannot get block at height %v", hh),
 			http.StatusBadRequest,
