+++ api/chains/chains.go
@@ -17,6 +17,7 @@ import (
 	"github.com/emerishq/demeris-api-server/api/apiutils"
 	"github.com/emerishq/demeris-api-server/api/database"
 	"github.com/emerishq/demeris-api-server/api/router/deps"
+	"github.com/emerishq/demeris-api-server/lib/apierrors"
 	"github.com/emerishq/demeris-api-server/lib/ginutils"
 	"github.com/emerishq/demeris-api-server/lib/stringcache"
 	"github.com/emerishq/demeris-api-server/sdkservice"
@@ -46,7 +47,7 @@ func GetChains(c *gin.Context) {
 	chains, err := d.Database.SimpleChains()
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve chains"),
 			http.StatusBadRequest,
@@ -127,7 +128,7 @@ func GetPrimaryChannelWithCounterparty(c *gin.Context) {
 	counterparty := c.Param("counterparty")
 	chain, err := d.Database.PrimaryChannelCounterparty(chainName, counterparty)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"primarychannel",
 			fmt.Errorf("cannot retrieve primary channel between %v and %v", chainName, counterparty),
 			http.StatusBadRequest,
@@ -173,7 +174,7 @@ func GetPrimaryChannels(c *gin.Context) {
 	chainName := c.Param("chain")
 	chain, err := d.Database.PrimaryChannels(chainName)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"primarychannel",
 			fmt.Errorf("cannot retrieve primary channels for %v", chainName),
 			http.StatusBadRequest,
@@ -283,7 +284,7 @@ func VerifyTrace(c *gin.Context) {
 
 		err = fmt.Errorf("cannot query list of chain ids, %w", err)
 
-		e := deps.NewError(
+		e := apierrors.New(
 			"denom/verify-trace",
 			fmt.Errorf("cannot query list of chain ids"),
 			http.StatusBadRequest,
@@ -372,7 +373,7 @@ func VerifyTrace(c *gin.Context) {
 
 				c.JSON(http.StatusOK, res)
 			} else {
-				e1 := deps.NewError(
+				e1 := apierrors.New(
 					"denom/verify-trace",
 					fmt.Errorf("failed querying for %s, error: %w", hash, err),
 					http.StatusBadRequest,
@@ -424,7 +425,7 @@ func VerifyTrace(c *gin.Context) {
 			return
 		}
 
-		e := deps.NewError(
+		e := apierrors.New(
 			"denom/verify-trace",
 			fmt.Errorf("database error, %w", err),
 			http.StatusInternalServerError,
@@ -451,7 +452,7 @@ func VerifyTrace(c *gin.Context) {
 
 	cbt, err := d.Database.ChainLastBlock(nextChain)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"denom/verify-trace",
 			fmt.Errorf("cannot retrieve chain status for %v", nextChain),
 			http.StatusInternalServerError,
@@ -584,7 +585,7 @@ func GetChainSupply(c *gin.Context) {
 	chain := ginutils.GetValue[cns.Chain](c, ChainContextKey)
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -613,7 +614,7 @@ func GetChainSupply(c *gin.Context) {
 
 	sdkRes, err := client.Supply(context.Background(), payload)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve supply from sdk-service"),
 			http.StatusBadRequest,
@@ -673,7 +674,7 @@ func GetDenomSupply(c *gin.Context) {
 
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -700,7 +701,7 @@ func GetDenomSupply(c *gin.Context) {
 		if len(sdkRes.Coins) != 1 {
 			cause = fmt.Errorf("expected 1 denom for chain: %s - denom: %s, found %v", chain.ChainName, denom, sdkRes.Coins)
 		}
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			cause,
 			http.StatusBadRequest,
@@ -739,7 +740,7 @@ func GetChainTx(c *gin.Context) {
 	chain := ginutils.GetValue[cns.Chain](c, ChainContextKey)
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -764,7 +765,7 @@ func GetChainTx(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve tx from sdk-service, %w", err),
 			http.StatusBadRequest,
@@ -804,7 +805,7 @@ func GetNumbersByAddress(c *gin.Context) {
 
 	resp, err := apiutils.FetchAccountNumbers(chainInfo, address)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"numbers",
 			fmt.Errorf("cannot retrieve account/sequence numbers for address %v", address),
 			http.StatusBadRequest,
@@ -844,7 +845,7 @@ func GetInflation(c *gin.Context) {
 
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -868,7 +869,7 @@ func GetInflation(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve inflation from sdk-service"),
 			http.StatusBadRequest,
@@ -906,7 +907,7 @@ func GetStakingParams(c *gin.Context) {
 
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -930,7 +931,7 @@ func GetStakingParams(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve staking params from sdk-service"),
 			http.StatusBadRequest,
@@ -968,7 +969,7 @@ func GetStakingPool(c *gin.Context) {
 
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -992,7 +993,7 @@ func GetStakingPool(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve staking pool from sdk-service"),
 			http.StatusBadRequest,
@@ -1029,7 +1030,7 @@ func GetMintParams(c *gin.Context) {
 	chain := ginutils.GetValue[cns.Chain](c, ChainContextKey)
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -1053,7 +1054,7 @@ func GetMintParams(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve mint params from sdk-service"),
 			http.StatusBadRequest,
@@ -1090,7 +1091,7 @@ func GetAnnualProvisions(c *gin.Context) {
 	chain := ginutils.GetValue[cns.Chain](c, ChainContextKey)
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -1114,7 +1115,7 @@ func GetAnnualProvisions(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve mint annual provision from sdk-service"),
 			http.StatusBadRequest,
@@ -1152,7 +1153,7 @@ func GetEpochProvisions(c *gin.Context) {
 
 	client, err := sdkservice.Client(chain.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 			http.StatusBadRequest,
@@ -1176,7 +1177,7 @@ func GetEpochProvisions(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve mint epoch provisions from sdk-service"),
 			http.StatusBadRequest,
@@ -1221,7 +1222,7 @@ func GetStakingAPR(c *gin.Context) {
 	)
 	aprString, err := aprCache.Get(c.Request.Context(), chainName)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot get APR"),
 			http.StatusBadRequest,
@@ -1242,7 +1243,7 @@ func GetStakingAPR(c *gin.Context) {
 
 	apr, err := strconv.ParseFloat(aprString, 64)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot convert apr to float"),
 			http.StatusBadRequest,
@@ -1273,7 +1274,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 		chain := ginutils.GetValue[cns.Chain](c, ChainContextKey)
 		client, err := sdkservice.Client(chain.MajorSDKVersion())
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot retrieve sdk-service for version %s with chain name %v", chain.CosmosSDKVersion, chain.ChainName),
 				http.StatusBadRequest,
@@ -1298,7 +1299,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 		})
 
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot retrieve staking pool from sdk-service"),
 				http.StatusBadRequest,
@@ -1320,7 +1321,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 		var stakingPoolData StakingPoolResponse
 		err = json.Unmarshal(stakingPoolRes.StakingPool, &stakingPoolData)
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot unmarshal staking pool"),
 				http.StatusBadRequest,
@@ -1341,7 +1342,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 
 		bondedTokens, err := strconv.Atoi(stakingPoolData.Pool.BondedTokens)
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot convert bonded_tokens to int"),
 				http.StatusBadRequest,
@@ -1366,7 +1367,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 		})
 
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot retrieve staking params from sdk-service"),
 				http.StatusBadRequest,
@@ -1388,7 +1389,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 		var stakingParamsData StakingParamsResponse
 		err = json.Unmarshal(stakingParamsRes.StakingParams, &stakingParamsData)
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot unmarshal staking params"),
 				http.StatusBadRequest,
@@ -1421,7 +1422,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 			if len(denomSupplyRes.Coins) != 1 {
 				cause = fmt.Errorf("expected 1 denom for chain: %s - denom: %s, found %v", chain.ChainName, bond_denom, denomSupplyRes.Coins)
 			}
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				cause,
 				http.StatusBadRequest,
@@ -1442,7 +1443,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 		// Hence, converting it to type coin to extract amount
 		coin, err := sdktypes.ParseCoinNormalized(denomSupplyRes.Coins[0].Amount)
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot convert amount to coin"),
 				http.StatusBadRequest,
@@ -1468,7 +1469,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 		})
 
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot retrieve inflation from sdk-service"),
 				http.StatusBadRequest,
@@ -1490,7 +1491,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 		var inflationData InflationResponse
 		err = json.Unmarshal(inflationRes.MintInflation, &inflationData)
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot unmarshal inflation"),
 				http.StatusBadRequest,
@@ -1511,7 +1512,7 @@ func getAPR(c *gin.Context) stringcache.HandlerFunc {
 
 		inflation, err := strconv.ParseFloat(inflationData.Inflation, 64)
 		if err != nil {
-			e := deps.NewError(
+			e := apierrors.New(
 				"chains",
 				fmt.Errorf("cannot convert inflation to float64"),
 				http.StatusBadRequest,
