+++ api/router/deps/deps.go
@@ -5,6 +5,7 @@ import (
 	kube "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/emerishq/demeris-api-server/api/database"
+	"github.com/emerishq/demeris-api-server/lib/apierrors"
 	"github.com/emerishq/demeris-api-server/lib/ginutils"
 	"github.com/emerishq/emeris-utils/logging"
 	"github.com/emerishq/emeris-utils/store"
@@ -37,7 +38,7 @@ func GetDeps(c *gin.Context) *Deps {
 }
 
 // WriteError logs and return client-facing errors
-func (d *Deps) WriteError(c *gin.Context, err Error, logMessage string, keyAndValues ...interface{}) {
+func (d *Deps) WriteError(c *gin.Context, err apierrors.Error, logMessage string, keyAndValues ...interface{}) {
 
 	// setting error id
 	value, ok := c.Request.Context().Value(logging.IntCorrelationIDName).(string)
