+++ api/tx/ibc.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/emerishq/demeris-api-server/api/router/deps"
+	"github.com/emerishq/demeris-api-server/lib/apierrors"
 	"github.com/emerishq/demeris-api-server/sdkservice"
 	sdkutilities "github.com/emerishq/sdk-service-meta/gen/sdk_utilities"
 	"github.com/gin-gonic/gin"
@@ -59,7 +60,7 @@ func GetDestTx(c *gin.Context) {
 
 	srcChainInfo, err := d.Database.Chain(srcChain)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve srcChainInfo with name %v", srcChain),
 			http.StatusBadRequest,
@@ -81,7 +82,7 @@ func GetDestTx(c *gin.Context) {
 	// validate destination srcChainInfo is present
 	destChainInfo, err := d.Database.Chain(destChain)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve srcChainInfo with name %v", destChain),
 			http.StatusBadRequest,
@@ -102,7 +103,7 @@ func GetDestTx(c *gin.Context) {
 
 	client, err := sdkservice.Client(srcChainInfo.MajorSDKVersion())
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve sdk-service for version %s with srcChainInfo name %v", srcChainInfo.CosmosSDKVersion, srcChainInfo.ChainName),
 			http.StatusBadRequest,
@@ -127,7 +128,7 @@ func GetDestTx(c *gin.Context) {
 	})
 
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve tx from sdk-service, %w", err),
 			http.StatusBadRequest,
@@ -155,7 +156,7 @@ func GetDestTx(c *gin.Context) {
 	// for now we just get the first seq number found and roll with it.
 	r := getIBCSeqFromTx(sdkRes)
 	if len(r) == 0 {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("provided transaction is not ibc transfer"),
 			http.StatusBadRequest,
@@ -186,7 +187,7 @@ func GetDestTx(c *gin.Context) {
 	// we're validating inputs and hence gosec-G107 can be ignored
 	resp, err := httpClient.Get(url) // nolint: gosec
 	if err != nil || resp.StatusCode != http.StatusOK {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve tx with packet sequence %s on %s", seqNum, destChain),
 			http.StatusBadRequest,
@@ -212,7 +213,7 @@ func GetDestTx(c *gin.Context) {
 
 	bz, err := io.ReadAll(resp.Body)
 	if err != nil {
-		e := deps.NewError(
+		e := apierrors.New(
 			"chains",
 			fmt.Errorf("cannot retrieve tx with packet sequence %s on %s", seqNum, destChain),
 			http.StatusBadRequest,
